+++ src/collective/contact/facetednav/browser/batchactions/manager.py
@@ -1,0 +1,11 @@
+from zope.interface import implements
+from zope.viewlet.interfaces import IViewletManager
+from zope.viewlet.manager import ConditionalViewletManager
+
+
+class IBatchActions(IViewletManager):
+    pass
+
+class BatchActionsViewletManager(ConditionalViewletManager):
+    implements(IBatchActions)
+