+++ src/collective/contact/facetednav/tests/test_setup.py
@@ -3,7 +3,9 @@
 
 from collective.contact.facetednav.testing import IntegrationTestCase
 from plone import api
-
+from plone.app.testing.helpers import login
+from plone.app.testing.interfaces import TEST_USER_NAME
+from eea.facetednavigation.interfaces import IPossibleFacetedNavigable
 
 class TestInstall(IntegrationTestCase):
     """Test installation of collective.contact.facetednav into Plone."""
@@ -16,6 +18,8 @@ class TestInstall(IntegrationTestCase):
     def test_product_installed(self):
         """Test if collective.contact.facetednav is installed with portal_quickinstaller."""
         self.assertTrue(self.installer.isProductInstalled('collective.contact.facetednav'))
+        self.assertTrue('mydirectory' in self.portal)
+        self.assertTrue(IPossibleFacetedNavigable.providedBy(self.portal.mydirectory))
 
     def test_uninstall(self):
         """Test if collective.contact.facetednav is cleanly uninstalled."""
@@ -28,3 +32,44 @@ class TestInstall(IntegrationTestCase):
         from collective.contact.facetednav.interfaces import ICollectiveContactFacetednavLayer
         from plone.browserlayer import utils
         self.failUnless(ICollectiveContactFacetednavLayer in utils.registered_layers())
+
+    def test_subtyper(self):
+        login(self.portal, TEST_USER_NAME)
+        directory = self.portal.mydirectory
+        subtyper = directory.unrestrictedTraverse('@@contact_faceted_subtyper')
+
+        subtyper.enable_select()
+        self.assertTrue(subtyper.can_select)
+        self.assertFalse(subtyper.can_enable_select())
+        self.assertTrue(subtyper.can_disable_select())
+        self.assertTrue(directory.unrestrictedTraverse('@@faceted_query').is_selectable())
+
+        subtyper.disable_select()
+        self.assertFalse(subtyper.can_select)
+        self.assertTrue(subtyper.can_enable_select())
+        self.assertFalse(subtyper.can_disable_select())
+        self.assertFalse(directory.unrestrictedTraverse('@@faceted_query').is_selectable())
+
+    def test_json_contacts(self):
+        login(self.portal, TEST_USER_NAME)
+        directory = self.portal.mydirectory
+        json_contacts = directory.unrestrictedTraverse('@@json-contacts')()
+        self.assertTrue(eval(json_contacts)[0].has_key('id'))
+
+        self.portal.REQUEST.form['type'] = 'organization'
+        json_contacts = directory.unrestrictedTraverse('@@json-contacts')()
+        self.assertEqual(len(eval(json_contacts)), 7)
+
+        self.portal.REQUEST.form['type'] = 'held_position'
+        json_contacts = directory.unrestrictedTraverse('@@json-contacts')()
+        self.assertEqual(len(eval(json_contacts)), 4)
+
+    def test_delete_action(self):
+        login(self.portal, TEST_USER_NAME)
+        directory = self.portal.mydirectory
+
+        self.assertIn('rambo', directory)
+        self.portal.REQUEST.form['uids'] = [directory.rambo.UID()]
+        delete_view = directory.unrestrictedTraverse('@@delete_selection')
+        delete_view()
+        self.assertNotIn('rambo', directory)